% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar-dois-numeros.R
\name{somar_dois_numero}
\alias{somar_dois_numero}
\title{Soma dois numeros}
\usage{
somar_dois_numero(x, y)
}
\arguments{
\item{x}{paramentro num 1}

\item{y}{parametro num 2}
}
\value{
soma
}
\description{
Soma dois numeros
}
\examples{
somar_dois_numero(1, 2)
}
